Remove unused functions and other cleanup.
Removed the function to get loaded fModules as it was both unused, and
didn't actually work. Additionally, variables that are defined by other modules are now checked for being
defined, and if they aren't, define them to make sure no error can
occur. The only exception is to sheliac.sh, as it can define the
variables BEFORE the other files define them.